22:15 (vibe code stage w7ch )

+++ Back-end/controllers/courseController.js
@@ -1,4 +1,11 @@
+const mongoose = require('mongoose');
 const Course = require('../models/Course');
+
+const httpError = (message, statusCode = 400) => {
+    const err = new Error(message);
+    err.statusCode = statusCode;
+    return err;
+};
 
 // @desc    Get all courses
 // @route   GET /api/courses
@@ -47,13 +54,14 @@ exports.getAllTopics = async (req, res, next) => {
 // @route   GET /api/courses/:id
 exports.getCourse = async (req, res, next) => {
     try {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            return next(httpError('Invalid course id', 400));
+        }
+
         const course = await Course.findById(req.params.id);
 
         if (!course) {
-            return res.status(404).json({
-                success: false,
-                error: 'Course not found'
-            });
+            return next(httpError('Course not found', 404));
         }
 
         res.status(200).json({
@@ -83,6 +91,10 @@ exports.createCourse = async (req, res, next) => {
 // @route   PUT /api/courses/:id
 exports.updateCourse = async (req, res, next) => {
     try {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            return next(httpError('Invalid course id', 400));
+        }
+
         const course = await Course.findByIdAndUpdate(
             req.params.id,
             req.body,
@@ -90,10 +102,7 @@ exports.updateCourse = async (req, res, next) => {
         );
 
         if (!course) {
-            return res.status(404).json({
-                success: false,
-                error: 'Course not found'
-            });
+            return next(httpError('Course not found', 404));
         }
 
         res.status(200).json({
@@ -109,13 +118,14 @@ exports.updateCourse = async (req, res, next) => {
 // @route   DELETE /api/courses/:id
 exports.deleteCourse = async (req, res, next) => {
     try {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            return next(httpError('Invalid course id', 400));
+        }
+
         const course = await Course.findByIdAndDelete(req.params.id);
 
         if (!course) {
-            return res.status(404).json({
-                success: false,
-                error: 'Course not found'
-            });
+            return next(httpError('Course not found', 404));
         }
 
         res.status(200).json({
@@ -131,13 +141,14 @@ exports.deleteCourse = async (req, res, next) => {
 // @route   POST /api/courses/:id/topics
 exports.addTopic = async (req, res, next) => {
     try {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            return next(httpError('Invalid course id', 400));
+        }
+
         const course = await Course.findById(req.params.id);
 
         if (!course) {
-            return res.status(404).json({
-                success: false,
-                error: 'Course not found'
-            });
+            return next(httpError('Course not found', 404));
         }
 
         course.topics.push(req.body);
@@ -179,22 +190,24 @@ exports.getCourseTopics = async (req, res, next) => {
 // @route   PUT /api/courses/:id/topics/:topicId
 exports.updateTopic = async (req, res, next) => {
     try {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            return next(httpError('Invalid course id', 400));
+        }
+
+        if (!mongoose.isValidObjectId(req.params.topicId)) {
+            return next(httpError('Invalid topic id', 400));
+        }
+
         const course = await Course.findById(req.params.id);
 
         if (!course) {
-            return res.status(404).json({
-                success: false,
-                error: 'Course not found'
-            });
+            return next(httpError('Course not found', 404));
         }
 
         const topic = course.topics.id(req.params.topicId);
 
         if (!topic) {
-            return res.status(404).json({
-                success: false,
-                error: 'Topic not found'
-            });
+            return next(httpError('Topic not found', 404));
         }
 
         Object.assign(topic, req.body);
@@ -213,16 +226,27 @@ exports.updateTopic = async (req, res, next) => {
 // @route   DELETE /api/courses/:id/topics/:topicId
 exports.deleteTopic = async (req, res, next) => {
     try {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            return next(httpError('Invalid course id', 400));
+        }
+
+        if (!mongoose.isValidObjectId(req.params.topicId)) {
+            return next(httpError('Invalid topic id', 400));
+        }
+
         const course = await Course.findById(req.params.id);
 
         if (!course) {
-            return res.status(404).json({
-                success: false,
-                error: 'Course not found'
-            });
+            return next(httpError('Course not found', 404));
         }
 
-        course.topics.pull(req.params.topicId);
+        const topic = course.topics.id(req.params.topicId);
+
+        if (!topic) {
+            return next(httpError('Topic not found', 404));
+        }
+
+        topic.deleteOne();
         await course.save();
 
         res.status(200).json({
